+++ exspy/_misc/eels/__init__.py
@@ -1,0 +1,5 @@
+from exspy._misc.eels.hydrogenic_gos import HydrogenicGOS
+from exspy._misc.eels.gosh_gos import GoshGOS
+from exspy._misc.eels.hartree_slater_gos import HartreeSlaterGOS
+
+__all__ = ["HydrogenicGOS", "GoshGOS", "HartreeSlaterGOS"]
