+++ exspy/_misc/eels/base_gos.py
@@ -22,7 +22,7 @@ import math
 import numpy as np
 from scipy import constants, integrate, interpolate
 
-from exspy.misc.elements import elements
+from exspy._misc.elements import elements
 from hyperspy.misc.export_dictionary import (
     export_to_dictionary,
     load_from_dictionary,
