+++ exspy/_misc/eels/effective_angle.py
