+++ exspy/_misc/eels/hartree_slater_gos.py
@@ -23,7 +23,7 @@ import numpy as np
 from scipy import constants
 
 from hyperspy.defaults_parser import preferences
-from exspy.misc.eels.base_gos import TabulatedGOS
+from exspy._misc.eels.base_gos import TabulatedGOS
 
 
 _logger = logging.getLogger(__name__)
