+++ exspy/_misc/eels/tools.py
@@ -25,7 +25,7 @@ import matplotlib.pyplot as plt
 from scipy import constants
 
 from hyperspy.misc.array_tools import rebin
-from exspy.misc.elements import elements as elements_db
+from exspy._misc.elements import elements as elements_db
 import hyperspy.defaults_parser
 
 _logger = logging.getLogger(__name__)
