--- exspy/misc/eels/__init__.py
@@ -1,5 +1,0 @@
-from exspy.misc.eels.hydrogenic_gos import HydrogenicGOS
-from exspy.misc.eels.gosh_gos import GoshGOS
-from exspy.misc.eels.hartree_slater_gos import HartreeSlaterGOS
-
-__all__ = ["HydrogenicGOS", "GoshGOS", "HartreeSlaterGOS"]
