+++ exspy/signals/eds.py
@@ -28,8 +28,8 @@ import hyperspy.api as hs
 from hyperspy import utils
 from hyperspy.signal import BaseSignal
 from hyperspy._signals.signal1d import Signal1D, LazySignal1D
-from exspy.misc.elements import elements as elements_db
-from exspy.misc.eds import utils as utils_eds
+from exspy._misc.elements import elements as elements_db
+from exspy._misc.eds import utils as utils_eds
 from hyperspy.misc.utils import isiterable
 from hyperspy.docstrings.plot import BASE_PLOT_DOCSTRING_PARAMETERS, PLOT1D_DOCSTRING
 from hyperspy.docstrings.signal import LAZYSIGNAL_DOC
