+++ exspy/signals/eels.py
@@ -47,9 +47,9 @@ from hyperspy.docstrings.signal import (
 )
 
 from exspy.docstrings.model import EELSMODEL_PARAMETERS
-from exspy.misc.elements import elements as elements_db
-from exspy.misc.eels.tools import get_edges_near_energy
-from exspy.misc.eels.electron_inelastic_mean_free_path import (
+from exspy._misc.elements import elements as elements_db
+from exspy._misc.eels.tools import get_edges_near_energy
+from exspy._misc.eels.electron_inelastic_mean_free_path import (
     iMFP_Iakoubovskii,
     iMFP_angular_correction,
 )
